requirements.txt
openai
pandas 
numpy
pillow
flask
#e .
